Rename `WoltLabSuite/Core/Component/File/File` to `WoltLabSuite/Core/Component/File/Helper`
