+++ ts/WoltLabSuite/Core/Component/File/Helper.ts
