+++ wcfsetup/install/files/js/WoltLabSuite/Core/Component/File/Helper.js
